+++ packages/wallet/src/initialization/instances/subscription-delegation-service/subscription-delegation-service.test.ts
@@ -52,9 +52,10 @@ describe('subscriptionDelegationService', () => {
         'ChompApiService:verifyDelegation',
         'ChompApiService:createIntents',
         'ChompApiService:getIntentsByAddress',
-        'ChompApiService:getServiceDetails',
         'DelegationController:signDelegation',
+        'MoneyAccountBalanceService:fetchBalanceWithFallback',
         'RemoteFeatureFlagController:getState',
+        'SubscriptionController:getPricing',
       ],
     });
   });